+++ app/src/main/kotlin/org/qownnotes/mobile/QOwnNotesApplication.kt
@@ -51,6 +51,7 @@ import org.qownnotes.mobile.data.MIGRATION_2_3
 import org.qownnotes.mobile.data.MIGRATION_3_4
 import org.qownnotes.mobile.data.MIGRATION_4_5
 import org.qownnotes.mobile.data.MIGRATION_5_6
+import org.qownnotes.mobile.data.MIGRATION_6_7
 import org.qownnotes.mobile.data.QOwnNotesDatabase
 import org.qownnotes.mobile.data.RoomAccountRepository
 import org.qownnotes.mobile.data.RoomNoteRepository
@@ -90,7 +91,8 @@ class ApplicationComponent(
                 MIGRATION_2_3,
                 MIGRATION_3_4,
                 MIGRATION_4_5,
-                MIGRATION_5_6
+                MIGRATION_5_6,
+                MIGRATION_6_7
             )
             .build(),
     private val backend: NoteBackend = NextcloudBackend(application),